+++ Core/Src/DisplayRegsiterModule/display_module.h
@@ -15,6 +15,6 @@
 #endif /* SRC_DISPLAYREGSITERMODULE_DISPLAY_MODULE_H_ */
 
 void sendSPIData(uint8_t data[], uint16_t byteSize);
-void DisplayInteger(uint8_t display);
+void DisplayInteger(uint16_t display);
 void _ConvertIntToChar(uint8_t display);
 void _pushCharToRegister(char rep);
